stage2/requirements.txt
altair==5.3.0
attrs==23.2.0
blinker==1.8.2
branca==0.7.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
docopt==0.6.2
folium==0.16.0
fonttools==4.43.1
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
importlib_resources==6.4.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.0
mdurl==0.1.2
mqtt-client==1.6.1
numpy==1.26.0
packaging==23.2
paho-mqtt==1.6.1
pandas==2.1.1
Pillow==10.0.1
plotly==5.17.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyarrow==16.0.0
pydeck==0.9.0
Pygments==2.18.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
referencing==0.35.1
requests==2.31.0
rich==13.7.1
rpds-py==0.18.1
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.30
streamlit==1.34.0
streamlit_folium==0.20.0
tenacity==8.2.3
terminaltables==3.1.10
toml==0.10.2
toolz==0.12.1
tornado==6.4
typing_extensions==4.11.0
tzdata==2023.3
urllib3==2.2.1
xyzservices==2024.4.0
zipp==3.18.1
